feat: add link to login page
