+++ src/components/Layout/Layout.scss
@@ -1,0 +1,37 @@
+.container {
+  min-height: calc(100vh - 105px);
+  position: relative;
+  box-sizing: border-box;
+  margin: 0 auto;
+} 
+
+
+header {
+  display: flex;
+  justify-content: space-between;
+  background-color: rgba(44, 53, 54, 0.3);
+  padding: 23px 0;
+  text-align: center;
+}
+
+header > div > a,
+header > div > i {
+  color: white;
+  text-decoration: none;
+  padding: 0 0.5rem;
+  transition: 0.5s;
+}
+
+header > div > a:hover {
+  color: #fff;
+  border-bottom: 4px solid #fff;
+  border-radius: 2px;
+  background-color: rgba(44, 53, 54, 0.2);
+}
+
+.footer-container {
+  padding: 6px;
+  color: #fff;
+  font-size: 12px;
+  background-color: rgba(44, 53, 54, 0.3);
+}
